use bootstrapcdn and google font cdn
using cloudfront for fonts isn't working out. There are too many issues
with CORS. With this setup the fonts from lab will still be loaded if a
developer is offline. This is because the css from google and bootstrap
will never be loaded, so the font definitions won't override the lab
font definitions.
[#70044336]