Update molecule forming model to use LJ, generalize for model #2
The hand-written calculation has been removed to use the model's
own LJ forces, which approximates the bond forces and energy
when scaled correctly.

Now that we're using real forces, the large repulsive force you can
obtain so dwarfs the attractive force that it only really makes sense
to make the "Repulsive" text large, to ovoid equating the two forces.

The one issue with this model is that, now that we're using real forces
to calculate a real PE, the model prevents you from dropping the atom
when the PE is positive (i.e. when they are close). We could make this
check optional in MD2D, perhaps.

[#74687984][#74688264]