Create Energy of Bond Formation interactive
This uses the now-generalized model from Forming a Molecule,
and uses the bond-length and energy values from
http://www.science.uwaterloo.ca/~cchieh/cact/c120/bondel.html

The space-fill views are sized such that bonding happens when
the views are overlapping by about 1/3, although this varies
for different combinations (O--O, with a long-range bond compared
to its radius, bonds slightly separated, which also nicely
indicates a weak bond).

One remaining question is what to do with the graph. There is
a reset button, allowing you to clear it, but I'm unsure if we
need to have six different colors representing all six
interactions. (Or if we need to clear the graph for every new pair,
but not doing that is nice because it allows direct comparisons.)


[#74688264]