Make viscosity dynamic to avoid changes in force vector
Previously, the force of friction would ramp up as the atoms accelerated.
This meant that the overall force on each atom shrunk in the first few
steps of the model, which was visually confusing (though correct). It also
created edge effects where the initial force at far distances was enough
to display a force vector, but once the atoms started moving the force would
be below the cutoff, causing the vectors to disappear.

This is fixed by removing friction from the model when the atoms are far
apart, and then gradually ramping it up as the atoms approach. 

This also allows us to tune the model to more quickly dampen the bouncing
than in the original model.

Time step decreased to preserve original model's speed.

[#83818508]