Rework of Zippies flying when air is heated with torches
This version starts with the torches unlit—just the air molecules are
moving—and has an animation of the torches flaring up that is triggered
by the user. The interactive is meant to be removed when the model is
places in the story. This zippier are pinned at the beginning out of
view and also in a electric field so that the air molecules don’t
accidentally get them flying.