Preparing for embedding in Flying Zippy story
All models will start() on load to show the action of the air
particles, but may have a function that has to be called by the html to
show the change action. Three of the models still have start action and
reload buttons. These need to come out on the next commit.

Changed the color of the Zippies to a darker purple.